==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year, month):
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    print(URL)
    r = requests.get(URL)
    if r.status_code == 200:
        path= f'../data/raw/rides_{year}-{month:02d}.parquet'
        with open(path, 'wb') as f:
            f.write(r.content)
            return path
    else:
        raise Exception(f'Error: {r.status_code}')


download_one_file_of_raw_data(2022, 1)

import pandas as pd
rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')
rides.head()


rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id',
}, inplace=True)

rides.head(10)

rides['pickup_datetime'].describe()

rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']
rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
